Move client config
